+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.3.0
+- **Breaking change:** Introduce `unique` option to remove duplicate values. Previously, duplicate values were removed by default, but now it's "opt in" for performance reasons.
+
 ## 0.2.0
-- **Breaking Change:** Default export `truthyStringsKeys`.
+- **Breaking change:** Default export `truthyStringsKeys`.